// src/app/shared/models/bank.ts
export class Bank {
  account_number: string;
  account_type: string;
  bank_name: string;
  bank_uuid: string;
  flag_primary: any;
  routing_number: string;
  stamp_created: any;
  stamp_created_by: {
    name: string;
    uuid: string;
  };
  stamp_updated: any;
  stamp_updated_by: {
    name: string;
    uuid: string;
  };
  user_uuid: string;
}